refactored tests into separate file
