+++ sktime/contrib/deeplearning_based/dl4tsc/mlp.py
@@ -20,7 +20,7 @@ import pandas as pd
 
 from sktime.utils.validation import check_X_y
 from sktime.contrib.deeplearning_based.basenetwork import BaseDeepLearner
-from sktime.contrib.deeplearning_based.basenetwork import test_network
+from sktime.classifiers.tests.test_dl4tscnetworks import test_network
 
 
 class MLP(BaseDeepLearner):
