+++ sktime/contrib/deeplearning_based/tuned_cnn.py
@@ -24,7 +24,7 @@ import numpy as np
 from sktime.utils.validation import check_X_y
 from sktime.classifiers.base import BaseClassifier
 from sktime.contrib.deeplearning_based.basenetwork import BaseDeepLearner
-from sktime.contrib.deeplearning_based.basenetwork import test_network
+from sktime.classifiers.tests.test_dl4tscnetworks import test_network
 
 from sklearn.model_selection import GridSearchCV
 from sklearn.model_selection import RandomizedSearchCV
